Move logger out of tools
