+++ parcels/fieldset.py
@@ -14,7 +14,7 @@ from parcels._core.utils.time import is_compatible as datetime_is_compatible
 from parcels._typing import Mesh
 from parcels.field import Field, VectorField
 from parcels.tools.converters import Geographic, GeographicPolar
-from parcels.tools.loggers import logger
+from parcels import logger
 from parcels.xgrid import _DEFAULT_XGCM_KWARGS, XGrid
 
 if TYPE_CHECKING:
