+++ parcels/tools/__init__.py
@@ -1,5 +1,5 @@
 from .converters import *
 from .exampledata_utils import *
 from .interpolation_utils import *
-from .loggers import *
+from .._logger import *
 from .statuscodes import *
